requirements.txt
fastapi==0.128.0
pyyaml==6.0.3
uvicorn==0.40.0